+++ src/inspect_ai/_eval/evalset.py
@@ -1,7 +1,7 @@
 import hashlib
 import logging
 from copy import deepcopy
-from typing import Any, Callable, Literal, NamedTuple, Set, cast
+from typing import Any, Literal, NamedTuple, Set, cast
 
 import rich
 from pydantic_core import to_json
@@ -17,6 +17,7 @@ from typing_extensions import Unpack
 
 from inspect_ai._util.error import PrerequisiteError
 from inspect_ai._util.file import basename, filesystem
+from inspect_ai._util.notgiven import NOT_GIVEN, NotGiven
 from inspect_ai.approval._policy import ApprovalPolicy
 from inspect_ai.log import EvalLog
 from inspect_ai.log._bundle import bundle_log_dir
@@ -34,11 +35,14 @@ from inspect_ai.model import (
 from inspect_ai.model._generate_config import GenerateConfig
 from inspect_ai.solver._solver import Solver, SolverSpec
 from inspect_ai.util import DisplayType, SandboxEnvironmentType
+from inspect_ai.util._display import init_display_type
 
 from .eval import eval, eval_init
-from .loader import ResolvedTask, resolve_task_args
-from .task import Epochs, Tasks
-from .task.task import PreviousTask, Task
+from .loader import resolve_task_args
+from .task import Epochs
+from .task.resolved import ResolvedTask
+from .task.task import PreviousTask
+from .task.tasks import Tasks
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +60,7 @@ def eval_set(
     retry_wait: float | None = None,
     retry_connections: float | None = None,
     retry_cleanup: bool | None = None,
-    model: str | Model | list[str] | list[Model] | None = None,
+    model: str | Model | list[str] | list[Model] | None | NotGiven = NOT_GIVEN,
     model_base_url: str | None = None,
     model_args: dict[str, Any] | str = dict(),
     task_args: dict[str, Any] | str = dict(),
@@ -107,9 +111,9 @@ def eval_set(
             (defaults to 0.5)
         retry_cleanup: Cleanup failed log files after retries
             (defaults to True)
-        model: Model(s) for
-            evaluation. If not specified use the value of the INSPECT_EVAL_MODEL
-            environment variable.
+        model: Model(s) for evaluation. If not specified use the value of the INSPECT_EVAL_MODEL
+            environment variable. Specify `None` to define no default model(s), which will
+            leave model usage entirely up to tasks.
         model_base_url: Base URL for communicating
             with the model API.
         model_args: Model creation args
@@ -154,7 +158,7 @@ def eval_set(
         max_samples: Maximum number of samples to run in parallel
             (default is max_connections)
         max_tasks: Maximum number of tasks to run in parallel
-            (default is 1)
+            (defaults to number of models being evaluated)
         max_subprocesses: Maximum number of subprocesses to
             run in parallel (default is os.cpu_count())
         max_sandboxes: Maximum number of sandboxes (per-provider)
@@ -177,13 +181,11 @@ def eval_set(
     """
 
     # helper function to run a set of evals
-    def run_eval(
-        tasks: list[Task] | list[PreviousTask], models: list[Model]
-    ) -> list[EvalLog]:
+    def run_eval(tasks: list[ResolvedTask] | list[PreviousTask]) -> list[EvalLog]:
         # run evals
         results = eval(
             tasks=tasks,
-            model=models,
+            model=None,  # ResolvedTask/PreviousTask already carries its model
             model_base_url=model_base_url,
             model_args=model_args,
             task_args=task_args,
@@ -231,30 +233,10 @@ def eval_set(
         # return results
         return results
 
-    # helper function to run a list of task groups
-    def run_task_groups(
-        task_groups: list[TaskGroup],
-        run_tasks: Callable[[list[ResolvedTask]], list[Task] | list[PreviousTask]],
-    ) -> list[EvalLog]:
-        logs: list[EvalLog] = []
-        for task_group in task_groups:
-            # alias
-            group_models, group_tasks = task_group
-
-            # info log
-            logger.info(
-                f"eval_set (running task group): {','.join([task.task.name for task in group_tasks])}: {group_models}"
-            )
-
-            # run the evals
-            logs.extend(
-                run_eval(
-                    tasks=run_tasks(group_tasks),
-                    models=group_models.models,
-                )
-            )
-
-        return logs
+    # initialise display (otherwise eval_init will set it to full)
+    display = init_display_type(display)
+    if display == "conversation":
+        raise RuntimeError("eval_set cannot be used with conversation display.")
 
     # resolve tasks
     models, _, resolved_tasks = eval_init(
@@ -331,15 +313,11 @@ def eval_set(
         pending_tasks = [
             task[1] for task in all_tasks if task[0] not in log_task_identifiers
         ]
-        task_groups = schedule_pending_tasks(pending_tasks)
 
         # we have some pending tasks yet to run, run them
-        if len(task_groups) > 0:
+        if len(pending_tasks) > 0:
             # run the tasks
-            run_logs = run_task_groups(
-                task_groups=task_groups,
-                run_tasks=lambda tasks: [task.task for task in tasks],
-            )
+            run_logs = run_eval(pending_tasks)
 
             # if this was the entire list of resolved tasks, return results
             if len(pending_tasks) == len(all_tasks):
@@ -365,42 +343,10 @@ def eval_set(
                     for task in resolved_tasks
                     if task_identifier(task) in failed_task_identifiers
                 ]
-                task_groups = schedule_retry_tasks(failed_tasks)
 
-                # execute task groups (run previous task so we get the samples from the log)
-                def run_previous_tasks(tasks: list[ResolvedTask]) -> list[PreviousTask]:
-                    def task_to_failed_log(task: ResolvedTask) -> Log:
-                        resolved_task_identifier = task_identifier(task)
-                        return next(
-                            log
-                            for log in failed_logs
-                            if log.task_identifier == resolved_task_identifier
-                        )
-
-                    previous_tasks: list[PreviousTask] = []
-                    for task, log in zip(tasks, map(task_to_failed_log, tasks)):
-                        # NOTE: we used to try to recreate registry objects by
-                        # by just passing the task name, but that didn't work
-                        # when evals were run from another directory. we may
-                        # want to bring this back but we'd need to resolve the
-                        # directory issues.
-
-                        # deepcopy so the same instance is not run twice
-                        prev_task = deepcopy(task.task)
-
-                        previous_tasks.append(
-                            PreviousTask(
-                                id=log.header.eval.task_id,
-                                task=prev_task,
-                                task_args=resolve_task_args(task.task),
-                                log=read_eval_log(log.info),
-                            )
-                        )
-
-                    return previous_tasks
-
-                retried_logs = run_task_groups(
-                    task_groups=task_groups, run_tasks=run_previous_tasks
+                # run previous tasks (no models passed b/c previous task already carries its model)
+                retried_logs = run_eval(
+                    tasks=as_previous_tasks(failed_tasks, failed_logs)
                 )
 
                 # return success
@@ -441,6 +387,42 @@ def eval_set(
 
     # return status + results
     return success, results
+
+
+# convert resolved tasks to previous tasks
+def as_previous_tasks(
+    tasks: list[ResolvedTask], failed_logs: list[Log]
+) -> list[PreviousTask]:
+    def task_to_failed_log(task: ResolvedTask) -> Log:
+        resolved_task_identifier = task_identifier(task)
+        return next(
+            log
+            for log in failed_logs
+            if log.task_identifier == resolved_task_identifier
+        )
+
+    previous_tasks: list[PreviousTask] = []
+    for task, log in zip(tasks, map(task_to_failed_log, tasks)):
+        # NOTE: we used to try to recreate registry objects by
+        # by just passing the task name, but that didn't work
+        # when evals were run from another directory. we may
+        # want to bring this back but we'd need to resolve the
+        # directory issues.
+
+        # deepcopy so the same instance is not run twice
+        prev_task = deepcopy(task.task)
+
+        previous_tasks.append(
+            PreviousTask(
+                id=log.header.eval.task_id,
+                task=prev_task,
+                task_args=resolve_task_args(task.task),
+                model=task.model,
+                log=read_eval_log(log.info),
+            )
+        )
+
+    return previous_tasks
 
 
 # filters to determine when we are done
@@ -574,7 +556,7 @@ def task_identifier(task: ResolvedTask | EvalLog) -> str:
         task_file = task.eval.task_file or ""
         task_name = task.eval.task
         task_args = task.eval.task_args
-        model = task.eval.model
+        model = str(task.eval.model)
 
     # hash for task args
     task_args_hash = hashlib.sha256(
@@ -615,80 +597,6 @@ class ModelList:
         model_names = [str(model) for model in self.models]
         model_names.sort()
         return ",".join(model_names)
-
-
-class TaskGroup(NamedTuple):
-    models: ModelList
-    tasks: list[ResolvedTask]
-
-
-# group into models => tasks for maximum parallelism
-def schedule_pending_tasks(pending_tasks: list[ResolvedTask]) -> list[TaskGroup]:
-    # build a map of task identifiers and the models they target
-    task_id_model_targets: dict[str, ModelList] = {}
-    for pending_task in pending_tasks:
-        task_id = task_identifier_without_model(task_identifier(pending_task))
-        if task_id not in task_id_model_targets:
-            task_id_model_targets[task_id] = ModelList([])
-        if pending_task.model not in task_id_model_targets[task_id].models:
-            task_id_model_targets[task_id].models.append(pending_task.model)
-
-    # build a list of unique model targets
-    unique_model_targets: Set[ModelList] = set(task_id_model_targets.values())
-
-    # create schedule
-    schedule: list[TaskGroup] = [
-        TaskGroup(models=model_target, tasks=[])
-        for model_target in unique_model_targets
-    ]
-
-    for models, tasks in schedule:
-        # which task ids have this set of models
-        task_ids: list[str] = []
-        for task_id, task_models in task_id_model_targets.items():
-            if task_models == models:
-                task_ids.append(task_id)
-
-        # find a task for each of these ids
-        for task_id in task_ids:
-            tasks.append(
-                next(
-                    (
-                        task
-                        for task in pending_tasks
-                        if task_id
-                        == task_identifier_without_model(task_identifier(task))
-                    )
-                )
-            )
-
-    # deterministic return order
-    schedule.sort(key=lambda x: str(x[0]))
-
-    return schedule
-
-
-# group into model => tasks (can't do multiple models b/c these are PreviousTask
-# instances (and therefore model/task pair specific -- we don't want to create
-# multiple instances of these tasks)
-def schedule_retry_tasks(retry_tasks: list[ResolvedTask]) -> list[TaskGroup]:
-    # build a list of unique model targets
-    unique_model_targets: Set[ModelList] = set()
-    for retry_task in retry_tasks:
-        unique_model_targets.add(ModelList([retry_task.model]))
-
-    # create a task group for reach model target
-    schedule: list[TaskGroup] = []
-    for model_target in unique_model_targets:
-        group_tasks = [
-            task for task in retry_tasks if ModelList([task.model]) == model_target
-        ]
-        schedule.append(TaskGroup(model_target, group_tasks))
-
-    # deterministic return order
-    schedule.sort(key=lambda x: str(x[0]))
-
-    return schedule
 
 
 def starting_max_connections(models: list[Model], config: GenerateConfig) -> int: